CMAQ 5.2.1 chemistry updates.
git-svn-id: https://svn.code.sf.net/p/olam-model/code/branches/dry_density@803 2bc74c4e-f164-4307-89b2-97abcff1156f

+++ CMAQ/CONST_DATA.f90
@@ -108,7 +108,7 @@ module const_data
 ! The NIST Reference on Constants, Units, and Uncertainty. US National
 ! Institute of Standards and Technology. June 2015. Retrieved 2017-04-21.
 ! http://physics.nist.gov/cgi-bin/cuu/Value?r
-  REAL, PARAMETER :: DRGASUNIV = 8.314459848D0
+  REAL( 8 ), PARAMETER :: DRGASUNIV = 8.314459848D0
 
   Real( 8 ), Parameter :: boltzmann = drgasuniv / davo  ! [ J/K ]
